+++ DockDoor/Views/Hover Window/Special Hover Windows/Calendar/CalendarView.swift
@@ -18,6 +18,7 @@ struct CalendarView: View {
     let bundleIdentifier: String
     let dockPosition: DockPosition
     let bestGuessMonitor: NSScreen
+    let dockItemElement: AXUIElement?
     let isEmbeddedMode: Bool
     let isPinnedMode: Bool
     let idealWidth: CGFloat?
@@ -32,6 +33,7 @@ struct CalendarView: View {
          bundleIdentifier: String,
          dockPosition: DockPosition,
          bestGuessMonitor: NSScreen,
+         dockItemElement: AXUIElement?,
          isEmbeddedMode: Bool = false,
          isPinnedMode: Bool = false,
          idealWidth: CGFloat? = nil,
@@ -41,6 +43,7 @@ struct CalendarView: View {
         self.bundleIdentifier = bundleIdentifier
         self.dockPosition = dockPosition
         self.bestGuessMonitor = bestGuessMonitor
+        self.dockItemElement = dockItemElement
         self.isEmbeddedMode = isEmbeddedMode
         self.isPinnedMode = isPinnedMode
         self.idealWidth = idealWidth
@@ -62,6 +65,7 @@ struct CalendarView: View {
                     bundleIdentifier: bundleIdentifier,
                     dockPosition: dockPosition,
                     bestGuessMonitor: bestGuessMonitor,
+                    dockItemElement: dockItemElement,
                     isPinnedMode: isPinnedMode,
                     appIcon: appIcon,
                     hoveringAppIcon: hoveringAppIcon,