+++ DockDoor/Views/Hover Window/Special Hover Windows/Media/MediaControlsFullView.swift
@@ -8,6 +8,7 @@ struct MediaControlsFullView: View {
     let bundleIdentifier: String
     let dockPosition: DockPosition
     let bestGuessMonitor: NSScreen
+    let dockItemElement: AXUIElement?
     let isPinnedMode: Bool
     @Binding var isArtworkExpandedFull: Bool
     @Binding var showingLyricsInFull: Bool
@@ -62,6 +63,7 @@ struct MediaControlsFullView: View {
                         WindowDismissalContainer(appName: appName,
                                                  bestGuessMonitor: bestGuessMonitor,
                                                  dockPosition: dockPosition,
+                                                 dockItemElement: dockItemElement,
                                                  minimizeAllWindowsCallback: { _ in })
                             .allowsHitTesting(false)
                     }