+++ DockDoor/Views/Settings/Help/BaseSettingsView.swift
@@ -12,7 +12,6 @@ struct BaseSettingsView<Content: View>: View {
             content
                 .padding(20)
         }
-        .frame(minWidth: 650, idealWidth: 700, maxWidth: .infinity,
-               minHeight: 675, idealHeight: 700, maxHeight: 700)
+        .frame(minWidth: 650, idealWidth: 700, minHeight: 650)
     }
 }